nixos/networkd: add Group Option to Link section for .network files
Interface Groups can be used to make role-based firewall configurations
an are useful to create static firewall rules for dynamic interfaces.

The range for this option was taken from the systemd.networkd man page:

Link groups are similar to port ranges found in managed switches. When
network interfaces are added to a numbered group, operations on all the
interfaces from that group can be performed at once. Takes an unsigned
integer in the range 0...2147483647. Defaults to unset.